examples/streamlit-memories/requirements.txt
# API Dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
httpx==0.24.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.12
psycopg2-binary>=2.9.9
alembic==1.10.4
pgvector==0.1.8

# Vector and embeddings
faiss-cpu>=1.7.4,<1.8.0
sentence-transformers>=2.2.2
numpy>=1.26.0,<2.0.0  # Pin numpy to 1.x to avoid compatibility issues with faiss

# LLM Integration 
openai==0.27.8

# UI
streamlit>=1.28.0
altair>=4.2.2
pandas>=2.1.0
pillow>=10.0.0  # For Python 3.13 compatibility

# Utilities
python-dateutil==2.8.2
tenacity==8.2.2
tqdm==4.65.0
loguru==0.7.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
requests-mock==1.10.0 